backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==2.0.2
gunicorn==20.1.0
pyserial==3.5
numpy==1.21.5
flask-socketio==5.1.1
# bluepy===1.3.0
flask-cors==3.0.10
joblib==1.1.0
eventlet==0.33.0
scikit-learn==0.24.2
